refactor: refactor paypal hooks and api calls
