+++ src/app/api/paypal/capture-order/route.ts
@@ -1,22 +1,46 @@
-import { paypalHooks } from "@/utils/paypal";
+import { useAccessToken } from "@/hooks/useAccessToken";
+import axios, { isAxiosError } from "axios";
 import { NextRequest, NextResponse } from "next/server";
 
+const PAYPAL_API = process.env.PAYPAL_API || 'https://api.sandbox.paypal.com'
 
 export async function POST(request: NextRequest){
   const {orderId} = await request.json()
-  const { captureOrder } = paypalHooks()
-
 
   try {
     const capture = await captureOrder(orderId)
     return NextResponse.json(capture)
-  } catch (error: any) {
-    console.error('Error in create-order route:', error.response?.data || error.message)
+  } catch (error) {
+    const message = isAxiosError(error)
+        ? error.response?.data || error.message
+        : error instanceof Error
+          ? error.message
+          : 'Unknown error';
+    console.error('Erro ao tentar criar order:', message)
     return NextResponse.json(
-      { error: 'Failed to capture order' },
+      { error: 'Erro ao tentar criar order' },
       { status: 500 }
     )
   }
 
+  async function captureOrder(orderId: string): Promise<any> {
+    const accessToken = await useAccessToken();
+    try {
+      const response = await axios({
+        method: 'post',
+        url: `${PAYPAL_API}/v2/checkout/orders/${orderId}/capture`,
+        headers: {
+          'Authorization': `Bearer ${accessToken}`,
+          'Content-Type': 'application/json'
+        }
+      });
+
+      return response.data;
+    } catch (error) {
+      console.error('Error capturing order:', error);
+      throw error;
+    }
+
+  }
   
 }
